backend/requirements.txt
fastapi
uvicorn
google-genai
icalendar
python-dotenv
